Update cause detail component tests
