--- src/components/CauseDetail/Tile/Tile.test.js
@@ -1,42 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter } from 'react-router-dom';
-
-// Redux imports
-// import { Provider } from 'react-redux';
-import { createStore, applyMiddleware, compose } from 'redux';
-import reduxThunk from 'redux-thunk';
-import reducers from '../../../../reducers/reducer';
-
-// Enzyme imports
-import Enzyme, { mount, shallow } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-
-// Component import
-import Tile from './Tile.js';
-
-Enzyme.configure({ adapter: new Adapter() });
-
-let store = createStore(
-  reducers,
-  compose(
-    applyMiddleware(reduxThunk)
-  )
-);
-
-const defaultProps = {
-  tileIcon: 'Africa',
-};
-
-const testElement =
-  <BrowserRouter>
-    <Tile {...defaultProps} store={store} />
-  </BrowserRouter>;
-
-describe('<Tile />', () => {
-
-  it('renders without crashing', () => {
-    const div = document.createElement('div');
-    ReactDOM.render(testElement, div);
-  });
-});
