+++ src/components/CauseDetail/TileSection/TileSection.test.js
@@ -1,42 +1,18 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter } from 'react-router-dom';
-
-// Redux imports
-// import { Provider } from 'react-redux';
-import { createStore, applyMiddleware, compose } from 'redux';
-import reduxThunk from 'redux-thunk';
-import reducers from 'reducers/reducer';
-
-// Enzyme imports
-import Enzyme, { mount, shallow } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
+import { shallow } from 'enzyme';
 
 // Component import
-import TileSection from './TileSection.js';
-
-Enzyme.configure({ adapter: new Adapter() });
-
-let store = createStore(
-  reducers,
-  compose(
-    applyMiddleware(reduxThunk)
-  )
-);
+import { TileSection } from './TileSection.js';
 
 const defaultProps = {
   match: { params: { id: 1 } },
 };
 
-const testElement =
-  <BrowserRouter>
-    <TileSection {...defaultProps} store={store} />
-  </BrowserRouter>;
+const wrapper = shallow(<TileSection {...defaultProps} />);
 
 describe('<TileSection />', () => {
 
   it('renders without crashing', () => {
-    const div = document.createElement('div');
-    ReactDOM.render(testElement, div);
+    expect(wrapper.exists('.TileSection')).toBe(true);
   });
 });
